Create cli package, separate curl parser/writer
